Remove dependency on `react-primitives`
and try to hack it with index.{platform} files instead!
