+++ fonts/native.ts
@@ -1,0 +1,9 @@
+import { Fonts } from "./types";
+
+const fonts: Fonts = {
+  MERRIWEATHER: 'Merriweather-Light',
+  CIRCULAR: 'CircularStd-Book',
+  SORAY: 'SoRay-ExtraBold',
+}
+
+export { fonts }
