+++ fonts/types.ts
@@ -1,0 +1,7 @@
+interface Fonts {
+  MERRIWEATHER: string,
+  CIRCULAR: string,
+  SORAY: string
+}
+
+export { Fonts }
