+++ fonts/web.ts
@@ -1,0 +1,9 @@
+import { Fonts } from "./types";
+
+const fonts: Fonts = {
+  MERRIWEATHER: 'Merriweather-Light',
+  CIRCULAR: 'CircularStd',
+  SORAY: 'SoRay'
+}
+
+export { fonts }
