+++ index.ts
@@ -1,5 +1,5 @@
 import { colors } from './colors'
-import { fonts } from './fonts'
+import { fonts } from './fonts/web'
 
 export {
   colors,
