+++ native.ts
@@ -1,0 +1,4 @@
+import { colors } from './colors'
+import { fonts } from './fonts/native'
+
+export { colors, fonts }
