+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@hedviginsurance/brand",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "description": "The hedvig brand package",
   "main": "dist/index.js",
   "repository": "git@github.com:HedvigInsurance/brand.git",
@@ -13,8 +13,5 @@
   "scripts": {
     "prepublishOnly": "tsc"
   },
-  "dependencies": {
-    "@types/react-primitives": "^0.6.2",
-    "react-primitives": "^0.6.1"
-  }
+  "dependencies": {}
 }
